Changed text on main page
Added excurse in history.
Flatten narrative to be clear for human who first time opens a site.
Avoided specific economic and computer words, and non-wikipedia links.
On this page we already have words 'usury', 'Bitcoin' and 'demurrage'.
We doing it for normal people, so text should be as simple as possible.