added link to community wiki
Added link to the FreiCoin Community Wiki run by Joe's Pool.

This wiki is available to be edited by approved editors. The only requirement to be approved as an editor is the desire to educate the community on crypto coins, alternative currencies and as a way for users to get howto's on various things such as how to set up a mining rig on various pools, how to get freicoin, and how to build freicoin.

rebasing commit